ref: move CreateFormModalContent related components into own folder
