+++ frontend/src/features/workspace/components/CreateFormModal/CreateFormModalContent/index.ts
@@ -1,0 +1,1 @@
+export { CreateFormModalContent } from './CreateFormModalContent'
